Split the wait/teardown rules out of the subprocess runner
`_subprocess_execution` carried `# pylint: disable=too-many-lines` at 441
lines, with a TODO pointing at issue #30 — which is closed, and was about
adding stdin support rather than about this refactor. The suppression was
being carried rather than earned, and the reference had rotted.

Extract `cuprum/_subprocess_wait.py`: `_cancel_pending_consumers`,
`_wait_for_exit_code`, `_drain_stream_consumers`, and
`_wait_for_exit_code_within_timeout`. That is a real seam rather than a cut to
hit a number — the runner module is about orchestration (spawn, wire streams,
assemble the result), while these four are the rules for *ending* a run: how a
deadline is applied, when the process is terminated, and how the consumers are
drained exactly once.

`_subprocess_execution` drops to 295 lines, so the suppression and the stale
TODO both go. Direct pylint now flags only `cuprum/sh.py` (701 lines), which
is pre-existing and untouched here.

Three details worth recording:

- `_spawn_stream_consumers` deliberately stays put. `test_safe_cmd_streams`
  patches it by module path, and moving it would have broken that silently.
- Importers are repointed at the new module rather than re-exported through
  the old one. A facade would have kept the diff smaller but left `__all__`
  advertising names the module no longer defines.
- The split necessarily lands on this branch, not the parent: the extracted
  code imports `_report_timeout_expiry` and `_report_teardown_drain_failure`
  from `_timeout_reporting`, which only exists here.

Pure move, no behaviour change: 955 Python tests pass, both modules score 10.00
on code health, and the `_SubprocessExecution` back-reference is
TYPE_CHECKING-only so no import cycle forms.
