+++ cuprum/unittests/test_subprocess_timeout_logging.py
@@ -15,7 +15,7 @@ import typing as typ
 import pytest
 
 from cuprum import _timeout_reporting
-from cuprum._subprocess_execution import (
+from cuprum._subprocess_wait import (
     _drain_stream_consumers,
     _wait_for_exit_code_within_timeout,
 )
