+++ cuprum/unittests/test_subprocess_timeout_observe.py
@@ -13,7 +13,7 @@ import typing as typ
 
 import pytest
 
-from cuprum._subprocess_execution import (
+from cuprum._subprocess_wait import (
     _drain_stream_consumers,
     _wait_for_exit_code_within_timeout,
 )
